ppp: avoid applying an iono slant factor twice.
ppp_res() was applying a slant factor to the iono delays that
already included the application of a slant factor, and this
caused the delays to be too great as the zenith angle increased.

The estimated iono delays are maintained as vertical delays,
and the slant factor is now applied to these in model_iono().