varerr: move towards being more consistent.
There are three variations of varerr(), necessarily different, but an
attempt has been made to move towards being more consistent. This
patch does not change the current implementations, the computed
variances should be the same as-is, rather it documents areas of
inconsistency and gives code options to make them more consistent.