nav eph: rework to index by the satellite number
The nav eph, geph, and seph arrays had been bundling in all the satellites in
the systems into one linear array, and users of these typically searched
linearly for entries with a matching satellite. For long post processed
computations this was a performance issue.

Rework these to be two layer arrays, firstly indexing by the satellite number,
and then by the entries for these particular satellites. This changes a linear
search for the matching satellites into an array index reference, which is a
large performance improvement.