filter: optimize temporary storage usage
The R input is not used after the call to filter() by any current
callers, so reuse it within filter() to avoid the allocation and
copying to a temporary array of the same size. If some future caller
wants R preserved then it will need to do so itself.

The computation in filter() can store back to the input x array
without the need for a temporary, so eliminate this temporary and make
the x argument both input and output.

Avoid a redundant copy back to the x and P inputs upon an error. These
were already preserved upon an error, this is just a minor performance
improvement.