rtk relpos: optimize compressed array usage
The Kalman filter() function was compressing arrays to speed the
computation. This has been hoisted into relpos() so that more use can
be made of this compressed array format. ddres() and a few other
functions in this path now work with some of these compressed arrays
directly and avoid the larger sized uncompressed arrays. This reduces
data cache usage and data copying which can give useful performance
improvements.