matmul: improve cache data access patterns and hoist tests.
Reorder the loop iteration over the arrays to prioritise sequential
linear access patterns, attempting to improving data cache efficiency.

Hoist the test for the special case of beta being zero out of the
loops. This is a necessary special case, not just for performance
optimization as it avoids reading from the C array. Split the case of
beta being no-zero into separate functions.

Then noting that all callers in the case of beta being non-zero have
beta being 1.0, so hard code this into the functions and remove the
beta argument. Further, noting that in this case the callers have
alpha being either 1.0 or -1.0, so split these into two separate
functions matmulp() and matmulm() respectively, and also remove the
alpha argument.

Noting that all callers in the case of beta being zero also have alpha
being 1.0, also remove the alpha argument from this specialized function.