kepler: continue on an iteration overflow
Continue to use the iterative solution in the exceptional case of an
iteration overflow, rather than returning with uninitialized function
outputs.

In this exceptional case, the solution is likely as good as can be
with the numerical precision.

The current return path was not setting the function outputs on that
return path and the callers may be using uninitialized data in this
case.