getbits*,setbits*: rewrite to better handle and report edge cases
Change the bit position and lengths to be unsigned values. This avoids
having to check if they are negative, and makes it clear that these
functions are only intended to be used with non-negative bit positions
and lengths.

Check and report bit lengths that are out of range.

Better handle some edge cases, such as lengths being zero.

Avoid left shits of unsigned integers which is 'undefined behaviour'.