rtksrv: have stats out streams respect the stats level
The rtkrcv command was not copying the system output options to both
of the output stream options, so the second streams always used the
initialization options.

Move the output of the stats residuals from outsolstat() to
rtkoutstat(), which allows the residuals to be output for the rtksvr
output streams when the out-outstat is set to 'residuals'. A new
argument is added to rtkoutstat() for the output level.