Rewrite source code style; pass buffer sizes and check accesses; use bool.
Rewrite the code style to reduce the scope of variables. The original
style had large variable definition blocks at the start of
functions. Moved on to C99 now so the scope be reduced. This generally
makes the code more readable, making it clearer where the scope of a
variable starts and often limiting the scope to a smaller block. Use
multiple instances of the same named variable if there are limited to
separate scopes, to make it clear that they are separate instances.

C99 supports the bool type, so make use of this for functions that
were returning just two numerical codes to represent a bool return
value. There remain functions returning a few numerical codes that
might better return an enum.

Pass buffer sizes along with buffer pointers more generally. Not all
instances are covered yet. Have the callee use this size to check for
and prevent buffer overflows. A range of string functions have been
added that check for buffer overflows, and ensure that strings are nul
terminated. A bounds check macro has been added.

Many instances of strcpy(), strncpt(), strcat(), sprintf() etc have
been replaced by safer alternatives. These make use of the newly
passed buffer sizes, or in many cases target a fixed size buffer known
to sizeof() at compile.

To better support bounds checking, and these new functions, code that
used pointers into buffers has been reworked to index off the buffer
base. There is likely a lot more to do on this.

Code that built up strings, appending to them, has often been reworked
to use a new string concatenation function that recalculates the
length. This is not fast path code so the overhead is trade off for a
simpler coding style, and a coding style that might better translate
to use a string library, or use a memory stream destination.

Many of these changes should better support a move to use a string
object, perhaps dynamic, perhaps reference counted. The buffer and
size pair would be replaced by an string object. Code written to
access the strings as a buffer base plus index could more easily be
reworked to index off a string object.