Long-double variant
This is intended to help development, where a little extra precision
can help to valid computations, to check if numerical precision is
affecting results, to check the numerical stability of algorithms etc.

Replace use of the float and double types with long-double.

Update uses of double constants to be long-double constants.

In a few instances constants have been updated for more precision.

Updating printf and scanf formats.

In a few instances algorithms have been updated to interate to a
greater precision.

The core libary and console commands have been updated and are
expected to work. There is no support for the QT or Windows GUI apps.

This is targeted at the Intel x86 80bit extended floating point
format, and has been build on Linux x86_64. Where it gives a modest
increase in numerical precision for a modest reduction in performance.

This is a work in progress.