decode_soltime: better handle the trailing separator
Two of the paths were blindly adding a separator length and could jump
the end of the string if it was short. The caller could attempt to
decode garbage beyond the end of the string.

The third path was not advancing over the time if there was no
trailing separator and was returning success in this case. The caller
could attempt to decode part of the time as other data, but would
surely fail gracefully.

Have these paths search for the trailing separator and return NULL if
not found, and if found then accurately return a pointer to the end of
this separator.