Rework the implementation of the position options
Add separate position option implementation states for llh and xyz
position types. The implementation of the antenna types was already
consistent with the configuration options for the antenna types, but
the implementation state in refpos and rovpos dropped the distinction
between llh and xyz position formats and so lost the distinction when
converting back to the configuation option.

Now the rovpos and refpos use the same set of states as the antenna
types, including llh and xyz position types, although the positions
stored in ru and bu are still always in xyz format.

This simplifies the option handling and little as there is now no
translation between the position types, but the positions are still
translated between xyz and llh as needed.

It is now possible to save antenna positions to configuration files in
ecef xyx format, whereas previously these would be saved in llh format
even if loaded in xyx format.

The fronts ends have been updated to handle the two position states
and to take advantage of the new xyz state. So that an antenna
position set in the GUI in xyz format now saves to a configuration
file in xyz format, rather than in llh format.

Also tried to address some issues with the propagation of this state,
both in the core options and front ends.