Qt gui setPath: rework, correct
Follow the logic in src/stream.c decodetcppath() to be consistent.

Update getPath() to only return those components available. It no
longer tacks on a ':' separator for the STR component is there is
none, or other unnecessary separators.

There might still be further issues with the presentation of the
STR component for ntrip servers and casters.