RINEX interface: define the system index
The RINEX code is using a particular index of systems and this is
exposed via the interface in the rnxopt_t mask, so define these in
rtklib.h along with the total number for the caller to rely on.

Make uses of these defines to avoid some uses of baked in constants.