convrnx: relax the test for duplicate data
Both convobs and convsbs were omitting data that was within the time
tolerance or later than the previous latest data time. This was
intended to omit some duplicate data. At least for the SBAS data this
heuristic was too strict as multiple messages typically occur with the
same time and messages and ephemeris were being omitted.

Relax the test to now keep data within the time tolerance.