Qt gui monitor: nav empty guard, widths, rtcm ssr
Rework the guarding of the nav eph. Use the per sat number of elements
to guard the offset index and if oob then use a zero element.

The width elements were short in setNavigation() and setSat() causing
OOB access. Just add the extra entries for now but the cause might
need review.

In showRtcmSsr(), slightly rework an iteration over the code indexes
to avoid CODE_NONE, although it was not broken and code2obs() does
catch an OOB code index.