rtcm3,rtcm3e: disable the carrier phase adjustments
For the messages 1001-1004 and 1009-1012, the step of 1500 cycles can
leave the carrier phase out of range of the representation, does not
appear to work as intended, as a smooth roll over, so just disable it
and encode the invalid value if out of range.

The carrier phase being out of range of the representation was not
being handled and is now caught and the invalid value is encoded.

The decoder attempt to smoothly roll over the carrier phase in steps
of 1500 cycles is also disabled, as encoders are not going to be
functional anyway, and it would introduce spurious jumps.