rtcm3e msm: rework rough range calculation, range checks
Rework the MSM rough range calculation, to consider the valid range
for the relevant measurements and choose a rough range that is
consistent with as many values possible while giving priority to
observations with both a pseudo range and carrier phase value.

Previously it had considered the first signal value and arbitrarily
chosen one value consistent with this, but there may well be values
consistent with the first value and also consistent with other values
which are now considered. This is an improvement when there are
multiple signals per satellite, a large set of values can be encoded.

A quick single pass pass greedy algorithm is used, but with multiple
concurrent search criteria which are then merged, so later values might
be out or range of their representation and emitted as invalid values.

The MSM pseudo range fine value components were being zeroed on a slip
or when out or range of their representation, which meant that the
receivers output was not being passed along as-is. With the above
improved selection of the rough range this is less of an issue, and
this is disabled by default, so that receivers out is passed through
as-is but pseudo range values out of range of their representation are
emitted as invalid values. The zeroing can be enabled using the new
-ZPHRNG option.