convrnx: lock time span pass for loss of lock determination
This pass recomputes the loss of lock indicator using the minimum lock
time provided by some of the observation input paths such as RTCM3 and
some of the raw decoders. It is useful as both a developer check for
comparison with real time loss of lock indicator, and as a post
processing pass to improve the loss of lock determination by using in
addition future lock times which are not available in real time
determination.