rtkpos ddidx restamb: rewrite for clarity
For ddidx(), replace the baked in striding over the state vector with
the use of the IB() macro. This makes the code less fragile if the
state vector layout changes and make it easier to explore alternative
double difference orderings.

Rewrite restamb() to use the index produced by ddidx() which it must
match, rather than striding over the state vector in the same manner
as ddidx(). Now many changes to the double difference indexing in
ddidx() do not necessarily require matching chances to restamb().

There remains some code in holdamb() striding over the state vector,
and that might need to match the indexing in ddidx(), and that is
still brittle to changes in ddidx().