rtkpos: rework antenna offsets, line of sight vectors per freq
Apply the antenna delta from the marker to ARP, and the per-frequency
PCO, before some of the other calculations such as the geometric
distance and line of sight vectors. The antenna PCV is later applied.
The line of sight vectors and now from the antenna phase center,
rather than from the marker position. This makes the calculations much
more robust to large antenna offsets. It was a trivial matter to also
use per-frequency line-of-sight vectors, so do that.

The azimuth and elevation are now also from the phase center rather
than the marker position, but still only per satellite.