pntpos: pass in the receiver index; and antpos fix
Pass in the receiver index, as a 'base' argument, which is 0 for the
rover and 1 for the base, to allow the respective options to be
selected rather than just using the rover options even for the
base. Start making use of this: in rescode() this is now used to
select the respective SNR and SNR mask.

For postpos.c where a one based receiver number and zero base index
were both be used, fix an instance in antpos() of indexing into option
anttype[] with an one base index which would be OOB for a value of
2. Try to clean up this code in this regard.