antpcv: split out the receiver PCV code from antmodel
Adding a new function antpcv(). Implement antmodel() using the
existing antpco() and the this new function to avoid code duplication,
with some repetition of code execution.

This is intended to allow the caller to apply the receiver PCO first
and then separately the PCV, for greater accuracy, just as these are
currently separated for satellites.