rtkrcv, rtknavi: support precise eph, clk, erp, and tide data
Predicted ephemeris, clock, and erp data can be used for RTK, so
support these.

For rtkrcv, allow these files to be specified on the command line, and
add the 'precdata' command to report the availability of the precise
data, and to allow loading updated data.

For rtknavi add options to specify these as input files, and report
the availability in the RTK monitor page.

Support loading tide data, the file-blqfile config option.