azimuth/elevation per receiver
Change the satellite status ssat_t azel to be per receiver, to better
handle long baseline RTK. This allows the elevation mask to be tested
at each receiver, and the obsdervation variance to be better estimated
for each receiver. For short baselines this is expected to make little
difference, but the satellite elevation can vary significantly for
long baselines and this is an input into selecting common satellites
and estimating observation variance.