rtkpos zdres: calculate for all signals even for the IFLC
This code had two separate paths: one path for the non-IFLC
calculation for each signal, and another path for the IFLC
calculation, and there was some duplication among these.

Have the IFLC result calculated from the results for each signal to
avoid this duplication.

This also helps prepare the code for the use of multiple LCs, and
reduces the maintenance burden.