options: expand to handle 6 frequency indices
Give an option for pos1-frequency to be up to 6 frequency indices,
adding values 5, and 6.

Expand the SNR masks to 6 frequency indices, up from 4. Rename the SNR
mask frequency options from pos1-snrmask_L{1,2,5} to and
pos1-snrmask{1,2,3,4,5,6}, so now referring to the frequency index
rather than attempting to refer to the signal.

Expand the eratio parameters to 6 frequency indices, up from
4. Reworking the option naming to refer to the signal index rather
than attempting to refer to the signal, so now
stats-eratio{1,2,3,4,5,6} for the respective frequency index.

Change pos1-frequency to be just an integer 1 to 6, and improve the
guard.