signal definition support, split BDS-2 and BDS-3
Add support to be able to redefine the mapping of signals to the
internal RTKLIB frequency index at runtime, and the respective code
priorities. Adding the pos1-sigdef config file option.

Add support for the console apps: rtkrcv using the config option; and
rnx2rtkp and convbin using the command line argument '--sigdef'.

Add support in the GUI apps rtknavi, rtkpost, and rtkconv, by adding
the 'Signal Definitions' edit box.

GUI apps rewrite the frequency dialog to dynamically fill the contents
and the also include the signal priorities and the frequency in MHz.

rtkrcv: add the 'frequencies' command to print a table of the signal
frequency index mapping and respective code priorities, and the
frequency in MHz, as for the gui apps frequencies dialog.

rtkconv: rework the code options selection to be dynamically filled,
so that these work with dynamically defined signal to frequency index
mappings.

Split BDS into BDS-2 and BDS-3, allowing each subsystem to have
separate signal to frequency index mappings and code
priorities. Solutions now also separate BDS with separate inter-system
time offsets and separate use in double differences.

Manage the decommissioning of BDS-2 and the move of BDS-3 satellites
to the lower PRNs. Add a variant of satsys() named satsyst() that
takes a time, and make use of this.