rtkrcv, rtknavi: support more solution input and out streams
Move towards the number of input and output streams being a
compilation option, adding the RTKSVRNIN, and RTKSVRNSOL defines and
also make better use of MAXSTRRTK rather than baking in
constants. Move the output streams to the end of the stream index
range so that adding more does not affect indices of the input and log
streams.

rtkrcv now builds with up to 5 input and 6 output streams and could
support more by simply adding move config options for them, and
rtknavi builds with 3 out streams.